requirements.txt
fastapi
loguru
pydantic
python-dotenv
python-youtube
sqlmodel
uvicorn[standard]
yt-dlp
